+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add startup event to initialize db.apply jobs
 - Update job_id after job submission
 - Fixed default event.uuid
+- Fixed atlas vector search
 
 #### New Features & Functionality
 
@@ -64,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the condition of the CDC job.
 - Fix form_template
 - Fix the duplicate children in the component.
+- Fix datatype for graph models
 
 ## [0.3.0](https://github.com/superduper-io/superduper/compare/0.3.0...0.2.0])    (2024-Jun-21)